Fixed the way test data is used in the Vertical leaderboard
If you want to test the Refresh feature, try it with the Vertical leaderboard! The larger data set should only show up if the user requests a refresh. Leaving and re-entering remakes the VeritcalLeaderboardFragment from scratch, which is why the smaller data set is queued up again in that case.